+++ demo/vue-app-new/package.json
@@ -9,6 +9,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@solana/web3.js": "^1.95.2",
     "@toruslabs/constants": "^14.0.0",
     "@toruslabs/openlogin-utils": "^8.2.1",
     "@toruslabs/vue-components": "^7.8.3",
@@ -16,11 +17,13 @@
     "@web3auth/base": "file:../../packages/base",
     "@web3auth/base-evm-adapter": "file:../../packages/adapters/base-evm-adapter",
     "@web3auth/base-solana-adapter": "file:../../packages/adapters/base-solana-adapter",
+    "@web3auth/coinbase-adapter": "file:../../packages/adapters/coinbase-adapter",
     "@web3auth/default-evm-adapter": "file:../../packages/adapters/default-evm-adapter",
     "@web3auth/default-solana-adapter": "file:../../packages/adapters/default-solana-adapter",
     "@web3auth/ethereum-provider": "file:../../packages/providers/ethereum-provider",
     "@web3auth/metamask-adapter": "file:../../packages/adapters/metamask-adapter",
     "@web3auth/modal": "file:../../packages/modal",
+    "@web3auth/modal-vue-composables": "file:../../packages/composables/modal-vue-composables",
     "@web3auth/no-modal": "file:../../packages/no-modal",
     "@web3auth/openlogin-adapter": "file:../../packages/adapters/openlogin-adapter",
     "@web3auth/phantom-adapter": "file:../../packages/adapters/phantom-adapter",
@@ -32,8 +35,6 @@
     "@web3auth/torus-solana-adapter": "file:../../packages/adapters/torus-solana-adapter",
     "@web3auth/wallet-connect-v2-adapter": "file:../../packages/adapters/wallet-connect-v2-adapter",
     "@web3auth/wallet-services-plugin": "^8.10.2",
-    "@web3auth/modal-vue-composables": "file:../../packages/composables/modal-vue-composables",
-    "@web3auth/coinbase-adapter": "file:../../packages/adapters/coinbase-adapter",
     "vue": "^3.4.31",
     "vue-i18n": "^9.13.1",
     "web3": "^4.11.1"
