+++ demo/vue-app-new/src/translations/en.json
@@ -26,6 +26,8 @@
     "isConnected": "Is connected ::::: {isConnected}",
     "isInitialized": "Is initialized ::::: {isInitialized}",
     "buttons": {
+      "btnClearConsole": "Clear Console",
+      "btnGetUserInfo": "Get User Info",
       "btnGetAccounts": "Get Accounts",
       "btnGetBalance": "Get Balance",
       "btnSendEth": "Send ETH",
